Self updater + More icon statuses
Had quite a hard time writing code to delete the old version of the
program. I ended up deleting my entire desktop in the process and
had to reprogram quite a bit(my source was in a folder on my
desktop). I removed the deletion code and it now has to be done
manually.